requirements.txt
pytest
pytest-cov
coverage
pandas
python-dotenv